Basic UI added with some styles added
I rewrote the JS to be nice and clean and streamline logic. I will
add a lockout to the buttons so that you have to start over rather than
keep going when it gets to a complete game. Other things added were
barebones styling. May add some images to make the game not so ugly
later on.